override keyword, Camera smoothing++, VAOs--
- use of C++11 override keyword
- oz
  * HashIndex::count & HashString::count fields removed, element count is
    already tracked by internal pool
- client
  * better camera smoothing
  * VAOs removed (cleaner code)
  * Colours class removed (lots of fields unused, others moved to Caelum)
- lua
  * ozCaelumSetRealTime()

+++ src/oz/arrays.hh
@@ -368,10 +368,10 @@ inline void aFree( Elem* aDest, int count )
  *
  * @ingroup oz
  */
-template <typename Elem, size_t SIZE>
+template <typename Elem, int SIZE>
 inline int aLength( const Elem ( & )[SIZE] )
 {
-  return int( SIZE );
+  return SIZE;
 }
 
 /**
